Legger til MeldekortDag
