+++ src/main/kotlin/no/nav/meldeplikt/meldekortservice/model/meldekortdetaljer/ords/Hode.kt
@@ -3,6 +3,8 @@ package no.nav.meldeplikt.meldekortservice.model.meldekortdetaljer.ords
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize
 import com.fasterxml.jackson.databind.annotation.JsonSerialize
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty
+import no.nav.meldeplikt.meldekortservice.model.meldekortdetaljer.ords.verdi.LongVerdi
+import no.nav.meldeplikt.meldekortservice.model.meldekortdetaljer.ords.verdi.StringVerdi
 import no.nav.meldeplikt.meldekortservice.utils.LocalDateDeserializer
 import no.nav.meldeplikt.meldekortservice.utils.LocalDateSerializer
 import java.time.LocalDate
@@ -28,14 +30,4 @@ data class Hode (
     @JsonSerialize(using = LocalDateSerializer::class)
     @JsonDeserialize(using = LocalDateDeserializer::class)
     val lestDato: LocalDate? = null
-)
-
-data class LongVerdi (
-    @JacksonXmlProperty(localName = "Verdi")
-    val verdi: Long
-)
-
-data class StringVerdi (
-    @JacksonXmlProperty(localName = "Verdi")
-    val verdi: String
 )
