+++ src/main/kotlin/no/nav/meldeplikt/meldekortservice/model/meldekortdetaljer/ords/verdi/IntVerdi.kt
@@ -1,0 +1,8 @@
+package no.nav.meldeplikt.meldekortservice.model.meldekortdetaljer.ords.verdi
+
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty
+
+data class IntVerdi (
+    @JacksonXmlProperty(localName = "Verdi")
+    val verdi: Int
+)
