+++ src/main/kotlin/no/nav/meldeplikt/meldekortservice/model/meldekortdetaljer/ords/verdi/StringVerdi.kt
@@ -1,0 +1,8 @@
+package no.nav.meldeplikt.meldekortservice.model.meldekortdetaljer.ords.verdi
+
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty
+
+data class StringVerdi (
+    @JacksonXmlProperty(localName = "Verdi")
+    val verdi: String
+)
